+++ config/servers.test.js
@@ -1,6 +1,3 @@
-// Server ports bind configuration
-// Each server is named server on specified address and port
-
 module.exports = {
   test: {
     protocol:  'http',
@@ -8,11 +5,6 @@ module.exports = {
     port:      8080,
     nagle:     true,
     slowTime:  '1s',
-    timeout:   '10s',
-    limit:     100,
-    limitHost: 50,
-    limitIP:   50,
-    limitSID:  50,
-    limitURL:  50
-  },
+    timeout:   '10s'
+  }
 };